modules/compute/eks/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
attrs==21.4.0
    # via
    #   cattrs
    #   jsii
aws-cdk-asset-awscli-v1==2.2.61
    # via aws-cdk-lib
aws-cdk-asset-kubectl-v20==2.1.1
    # via aws-cdk-lib
aws-cdk-asset-node-proxy-agent-v5==2.0.50
    # via aws-cdk-lib
aws-cdk-lambda-layer-kubectl-v23==2.0.3
    # via -r requirements.in
aws-cdk-lib==2.51.0
    # via
    #   -r requirements.in
    #   aws-cdk-lambda-layer-kubectl-v23
    #   cdk-nag
boto3==1.26.81
    # via
    #   -r requirements.in
    #   moto
botocore==1.29.92
    # via
    #   boto3
    #   moto
    #   s3transfer
cattrs==22.1.0
    # via jsii
cdk-nag==2.12.29
    # via -r requirements.in
certifi==2023.7.22
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.1.0
    # via requests
constructs==10.0.91
    # via
    #   -r requirements.in
    #   aws-cdk-lambda-layer-kubectl-v23
    #   aws-cdk-lib
    #   cdk-nag
cryptography==41.0.4
    # via moto
deepmerge==1.1.0
    # via -r requirements.in
idna==3.4
    # via requests
jinja2==3.1.2
    # via moto
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsii==1.74.0
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-lambda-layer-kubectl-v23
    #   aws-cdk-lib
    #   cdk-nag
    #   constructs
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
moto==4.1.11
    # via -r requirements.in
packaging==23.0
    # via -r requirements.in
publication==0.0.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-lambda-layer-kubectl-v23
    #   aws-cdk-lib
    #   cdk-nag
    #   constructs
    #   jsii
pycparser==2.21
    # via cffi
python-dateutil==2.8.2
    # via
    #   botocore
    #   jsii
    #   moto
pyyaml==6.0
    # via
    #   -r requirements.in
    #   responses
requests==2.31.0
    # via
    #   moto
    #   responses
responses==0.23.1
    # via moto
s3transfer==0.6.0
    # via boto3
six==1.16.0
    # via python-dateutil
typeguard==2.13.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-lambda-layer-kubectl-v23
    #   aws-cdk-lib
    #   jsii
types-pyyaml==6.0.12.10
    # via responses
typing-extensions==4.2.0
    # via jsii
urllib3==1.26.18
    # via
    #   botocore
    #   requests
    #   responses
werkzeug==2.3.6
    # via moto
xmltodict==0.13.0
    # via moto
